# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lexora Vocabulary Learning API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Word {
  id: number;
  word: string;
  meaning: string;
  partOfSpeech: string;
}

export type ExampleType = (typeof ExampleType)[keyof typeof ExampleType];

export const ExampleType = {
  casual: "casual",
  professional: "professional",
} as const;

export interface Example {
  id: number;
  wordId: number;
  type: ExampleType;
  sentence: string;
}

export type UserWordStatusStatus =
  (typeof UserWordStatusStatus)[keyof typeof UserWordStatusStatus];

export const UserWordStatusStatus = {
  known: "known",
  unknown: "unknown",
} as const;

export interface UserWordStatus {
  wordId: number;
  status: UserWordStatusStatus;
  updatedAt: string;
}

export type WordWithStatusStatus =
  | (typeof WordWithStatusStatus)[keyof typeof WordWithStatusStatus]
  | null;

export const WordWithStatusStatus = {
  known: "known",
  unknown: "unknown",
} as const;

export interface WordWithStatus {
  id: number;
  word: string;
  meaning: string;
  partOfSpeech: string;
  status?: WordWithStatusStatus;
}

export type WordDetailStatus =
  | (typeof WordDetailStatus)[keyof typeof WordDetailStatus]
  | null;

export const WordDetailStatus = {
  known: "known",
  unknown: "unknown",
} as const;

export interface WordDetail {
  id: number;
  word: string;
  meaning: string;
  partOfSpeech: string;
  examples: Example[];
  status?: WordDetailStatus;
}

export type SwipeBodyStatus =
  (typeof SwipeBodyStatus)[keyof typeof SwipeBodyStatus];

export const SwipeBodyStatus = {
  known: "known",
  unknown: "unknown",
} as const;

export interface SwipeBody {
  status: SwipeBodyStatus;
}

export type UpdateWordStatusBodyStatus =
  (typeof UpdateWordStatusBodyStatus)[keyof typeof UpdateWordStatusBodyStatus];

export const UpdateWordStatusBodyStatus = {
  known: "known",
  unknown: "unknown",
} as const;

export interface UpdateWordStatusBody {
  status: UpdateWordStatusBodyStatus;
}

export interface DailyLesson {
  words: WordWithStatus[];
  date: string;
}

export interface UserStats {
  totalWords: number;
  knownWords: number;
  unknownWords: number;
  unstudiedWords: number;
  streakDays: number;
}

export interface ErrorResponse {
  error: string;
}

export type ListWordsParams = {
  /**
   * Search query for filtering words
   */
  search?: string;
  /**
   * Filter by word status
   */
  status?: ListWordsStatus;
};

export type ListWordsStatus =
  (typeof ListWordsStatus)[keyof typeof ListWordsStatus];

export const ListWordsStatus = {
  known: "known",
  unknown: "unknown",
  all: "all",
} as const;