# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lexora Vocabulary Learning API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all words
 */
export const ListWordsQueryParams = zod.object({
  search: zod.coerce
    .string()
    .optional()
    .describe("Search query for filtering words"),
  status: zod
    .enum(["known", "unknown", "all"])
    .optional()
    .describe("Filter by word status"),
});

export const ListWordsResponseItem = zod.object({
  id: zod.number(),
  word: zod.string(),
  meaning: zod.string(),
  partOfSpeech: zod.string(),
  status: zod.enum(["known", "unknown"]).nullish(),
});
export const ListWordsResponse = zod.array(ListWordsResponseItem);

/**
 * @summary Get word detail
 */
export const GetWordParams = zod.object({
  id: zod.coerce.number(),
});

export const GetWordResponse = zod.object({
  id: zod.number(),
  word: zod.string(),
  meaning: zod.string(),
  partOfSpeech: zod.string(),
  examples: zod.array(
    zod.object({
      id: zod.number(),
      wordId: zod.number(),
      type: zod.enum(["casual", "professional"]),
      sentence: zod.string(),
    }),
  ),
  status: zod.enum(["known", "unknown"]).nullish(),
});

/**
 * Returns words prioritizing unknown words, with known words appearing less frequently
 * @summary Get words for swipe session
 */
export const GetSwipeQueueResponseItem = zod.object({
  id: zod.number(),
  word: zod.string(),
  meaning: zod.string(),
  partOfSpeech: zod.string(),
  status: zod.enum(["known", "unknown"]).nullish(),
});
export const GetSwipeQueueResponse = zod.array(GetSwipeQueueResponseItem);

/**
 * @summary Record a swipe decision
 */
export const RecordSwipeParams = zod.object({
  wordId: zod.coerce.number(),
});

export const RecordSwipeBody = zod.object({
  status: zod.enum(["known", "unknown"]),
});

export const RecordSwipeResponse = zod.object({
  wordId: zod.number(),
  status: zod.enum(["known", "unknown"]),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Toggle word known/unknown status
 */
export const UpdateWordStatusParams = zod.object({
  wordId: zod.coerce.number(),
});

export const UpdateWordStatusBody = zod.object({
  status: zod.enum(["known", "unknown"]),
});

export const UpdateWordStatusResponse = zod.object({
  wordId: zod.number(),
  status: zod.enum(["known", "unknown"]),
  updatedAt: zod.coerce.date(),
});

/**
 * Returns 3-5 words prioritizing unknown words for the daily lesson
 * @summary Get today's lesson words
 */
export const GetDailyLessonResponse = zod.object({
  words: zod.array(
    zod.object({
      id: zod.number(),
      word: zod.string(),
      meaning: zod.string(),
      partOfSpeech: zod.string(),
      status: zod.enum(["known", "unknown"]).nullish(),
    }),
  ),
  date: zod.coerce.date(),
});

/**
 * @summary Get user learning statistics
 */
export const GetStatsResponse = zod.object({
  totalWords: zod.number(),
  knownWords: zod.number(),
  unknownWords: zod.number(),
  unstudiedWords: zod.number(),
  streakDays: zod.number(),
});