# lib/api-zod/src/generated/types/dailyLesson.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lexora Vocabulary Learning API
 * OpenAPI spec version: 0.1.0
 */
import type { WordWithStatus } from "./wordWithStatus";

export interface DailyLesson {
  words: WordWithStatus[];
  date: Date;
}