# lib/api-zod/src/generated/types/example.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lexora Vocabulary Learning API
 * OpenAPI spec version: 0.1.0
 */
import type { ExampleType } from "./exampleType";

export interface Example {
  id: number;
  wordId: number;
  type: ExampleType;
  sentence: string;
}