# lib/api-zod/src/generated/types/swipeBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lexora Vocabulary Learning API
 * OpenAPI spec version: 0.1.0
 */
import type { SwipeBodyStatus } from "./swipeBodyStatus";

export interface SwipeBody {
  status: SwipeBodyStatus;
}