# lib/api-zod/src/generated/types/updateWordStatusBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lexora Vocabulary Learning API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateWordStatusBodyStatus } from "./updateWordStatusBodyStatus";

export interface UpdateWordStatusBody {
  status: UpdateWordStatusBodyStatus;
}