# lib/api-zod/src/generated/types/userStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lexora Vocabulary Learning API
 * OpenAPI spec version: 0.1.0
 */

export interface UserStats {
  totalWords: number;
  knownWords: number;
  unknownWords: number;
  unstudiedWords: number;
  streakDays: number;
}