# lib/api-zod/src/generated/types/userWordStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lexora Vocabulary Learning API
 * OpenAPI spec version: 0.1.0
 */
import type { UserWordStatusStatus } from "./userWordStatusStatus";

export interface UserWordStatus {
  wordId: number;
  status: UserWordStatusStatus;
  updatedAt: Date;
}