# lib/api-zod/src/generated/types/word.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lexora Vocabulary Learning API
 * OpenAPI spec version: 0.1.0
 */

export interface Word {
  id: number;
  word: string;
  meaning: string;
  partOfSpeech: string;
}