# lib/api-zod/src/generated/types/wordDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lexora Vocabulary Learning API
 * OpenAPI spec version: 0.1.0
 */
import type { Example } from "./example";
import type { WordDetailStatus } from "./wordDetailStatus";

export interface WordDetail {
  id: number;
  word: string;
  meaning: string;
  partOfSpeech: string;
  examples: Example[];
  status?: WordDetailStatus;
}